Adding missing validation

+++ app/src/main/java/com/neo4j/pregel/FrequentSubgraphMiningPregel.java
@@ -68,7 +68,7 @@ public class FrequentSubgraphMiningPregel implements PregelComputation<FrequentS
 
             for (int i = 0; i < new_fsms.length; i++) {
                 if (i == new_fsms.length - 1) {
-                    new_fsms[i] = msg;
+                    new_fsms[i] = (long) msg;
                 } else {
                     new_fsms[i] = fsms[i];
                 }
